Use Jackson to deserialize Configuration objects.
SnakeYAML is really not very good at object mapping, and Jackson is
awesome at that. So I use SnakeYAML to parse the YAML, build up a
JSON equivalent object, and then parse that with Jackson. It's not
exactly straight-forward, but it's easier to do than write my own
YAML parser for Jackson.