Some formatting changes.
Also removed the warnings about setting the servlet and filter names. Setting the name after calling setHeldClass (which is called in the ServletHolder/FilterHolder constructors when passing a class instead of an instance) simply overwrites the default name, which is the canonical class name and the class hashCode. Given that overwriting the default name is the intended behavior, the warnings seemed a little dire.