Fixing a build issue in Scala example
To be honest I have absoloutely no idea what I am doing. I watched a
video by @tnm at Basho, played around with maven and tried the "Getting
Started" section of dropwizard's manual. It was in Java but @tnm was
talking about Scala. I searched and found this example in dropwizard's
repo, and tried run it using "mvn scala:run", but it just failed. Made
this changes, and it started working.