Add Maven Archetypes for skeleton projects
It's a good idea to have a template for new projects to help users
adhere to best practices for project layout.

Provides:

  * A Java Archetype, `java-simple`, for creating new Java Dropwizard
  projects.
  * A Scala Archetype, `scala-simple`, for creating new Scala Dropwizard
  projects.
  * A utility, `dropwizard-create`, for bootstrapping a new project
  using the Archetypes.

The POM for generated projects will:

  * Generate JavaDoc/ScalaDoc, and build an attached doc JAR.
  * Build and attach a source JAR.
  * Build shaded JAR (optional)
  * Build an executable JAR for the Service.
  * Generate Maven Site

The `dropwizard-create` utility is provided so that users don't have to
grapple with Mavens' appalling UI for generating projects from
Archetypes. It ensures that all the necessary properties are provided to
the Archetypes and provides some sensible defaults for the lazy.