Fix IntelliJ trying to parse archetype templates
IntelliJ automatically adds the `resources` of a project as a content
root, making it include the `*.java` and `*.scala` template files of
each archetype.

Since they're full of velocity template variables, they fail to compile
and IntelliJ complains.

To fix this, we introduce a new required property: `language`, which is
used as the file extension for these templates when the project is
generated.

The default value of this property is defined by the archetype itself,
so it's not necessary to define it when generating the project.