Fixed an issue which prevent from using indexFile when the assets are mapped on "/" and the service is started from JAR file :
The path to the indexFile was built this way :  absoluteRequestedResourcePath + '/' + indexFile
When the assets are mapped on "/", this leads to : "my/resource/path/" + '/' + "myindex.html".
And when trying to get a resource url with a double slash (//) in path, it works from filesystem ("exploded jar"), when starting the service from Eclipse for example, but not within a JAR.

The fix simply trims "/" from absoluteRequestedResourcePath to make sure we never get a double slash.