Add '--configtest' parameter to ConfiguredCommand
The optional '--configtest' (or '-T') command line argument runs the command
until the configuration file has been parsed and validated and quits the command
after that. This is useful for testing the validity of a configuration file
before a service is being restarted and could be used in init scripts.