Provide Duration APIs as well as (long, TimeUnit)
Several components require settings a duration (for timeouts etc.) that
require a `long` and `TimeUnit` pair. Lets provide an API that accepts a
`Duration` instead, as it's often less cumbersome to work with,
especially when working with a `Configuration` that already provides the
value as a `Duration`.

Note: this adds `dropwizard-util` as a dependency of
`dropwizard-lifecycle`.