Add %app and %pid patterns for syslog appenders.
When present (including by default) the tokens `%app` and `%pid` will be
replaced with the name of the Dropwizard application and PID of the
process respectively.

This is in line with the convention for many Linux processes to prefix
their log messages with: "ApplicationName[PID]: ".

The PID can only be determined automatically for some UNIX JVMs
(including HotSpot). For others, it will be omitted entirely.