Make Liquibase migrations file an optional CLI arg
Sometimes it's desirable to maintain/package your database migrations
outside of your application JAR (e.g. when the migrations file is
massive).

Lets add an optional argument to all Liquibase Commands that permits
specifying the path to the migrations file in the local filesystem.

Omitting this option retains the default behaviour.

Note: I'm not entirely sure how best to test Command arguments, so there
are no tests. I used `dropwizard-example` to check that it roughly works
as expected, but that doesn't really qualify as testing.