Publish GitHub Pages during `site-deploy` goal
This was rather painful. First one has to generate the aggregated site for
all sub modules by running `mvn site:site site:stage` with the parent project's
POM (ie. not inside a subdirectory) after which one can deploy the result
(which resides in "target/staging") to GitHub Pages by executing the `site-deploy`
goal for the parent project.

The distributionManagement.site stuff is required for the maven-site-plugin to be
happy and allow the execution of `site:stage` but it should not deploy to the
URL give in distributionManagement.site.url using Wagon which is why skipDeploy
has to be set to true.

To put it short: ┻━┻ ︵ヽ(`Д´)ﾉ︵ ┻━┻