Quote applicationName in SyslogAppenderFactory
When using a user-defined String for the replacement of a regular
expression, we should be quoting it with `Matcher.quoteReplacement`, in
case they use any reserved characters (`/` or `$`).

In this case, we ran in to a problem with a Scala application with a
singleton object for the application, which gets compiled to a class
with a trailing `$`, causing a `StringIndexOutOfBoundsException` here.