Fix csv metrics documentation. Directory creation
When the Metric's CSV reporter is enabled, it doesn't write to a file, but
instead writes a series of files to a specified directory. This commit fixes
all references of the single file notion.

This commit also creates the CSV reporter directory if it doesn't already
exist on start up. An appropriate test was created to confirm behavior.