Refactoring logEndpoints, filter by Resource.from() != null instead of annotationPresent(Path.class)
This fixes #726, with 0.8.0 separating REST-interface and implementing bean was not supported by the configuration log since the implementing bean itself is not annotated with @Path.
While working on this issue, I removed a lot of duplicate lines dealing with merging and filtering getClasses() and getSingletons(). Also, everything related to Endpoint-Logging moved into a inner class to isolate concerns.