Configure Jersey thread pool with a bounded queue
Jersey's async thread pool is configured with 2 parameters:
*minThread* and *maxThreads*. The problem is that the pool
will not grow up than *minThread* because its work queue is
unbounded. It's a property of JDK ThreadPoolExecutor.

A solution to this problem is to set a bounded queue as the
work queue and provide facility to users to set the queue
size.