Add facility to override the default configuration
Fix #636, Fix #865

If a YAML configuration file is not provided, Dropwizard by default
uses an empty object tree as a config representation.

This approach has a flaw in case of configuration overriding.

For this to work, a full object tree analogous by structure with
the configuration should be present.

A better approach is to instantiate a configuration object with defaults,
ask Jackson to build an object tree and pass this tree to the overriding
routine. After that Jackson will create a desired configuration
from a changed tree.

In case if there is no overriding Jackson will just build the same
configuration object as the default.