Add Jersey DropwizardApacheConnector
Add our own version that uses the pre-configured Apache HTTP client.

Dropwizard has a stable and mature Apache HTTP client with an
extensive configuration. Rather than trying to simulate configuration
with Jersey properties, we can just use a built Apache client and
write a custom connector that does a glue work between Jersey and
Apache client's abstractions.

With this approach we have a single place of configuration.
That reduces burden of supporting a new feature both in Apache and
Jersey clients.