Support for a custom name of a Hibernate bundle
Resolve #855

By default the name of a Hibernate bundle (actually "hibernate)" is
hardcoded. This name is used for the database pool metrics and the
health check associated with the bundle. This causes conflicts when
several bundles used alongside in a project.

This change adds to the HibernateBundle class a protected method which
returns a bundle name. This method can be overridden by users who want
to set a custom name. The new name will be used for the bundle's
database pool metrics and health check.