Change filtering behaviour when excludes and includes are present
Resolve #890, Close #895

When the includes and excludes rules are present, we should change
metrics filtering behaviour to more obvious.

Previously, when both these rules were present, the excludes rule
had the following meaning: exclude these metrics from all metrics
+ include those from the includes list.

This is not what users expect, when they start with the inclusion
list: suddenly new metrics are reported which was not specified in
any list.

More anticipated behaviour is to treat the exclusion list over the
inclusion list. By default the inclusion list is empty (all metrics
are reported). Then if the inclusion list is changed, then the
exclusion list excludes metrics from specified inclusion rules,
but not all metrics.

Thus filtering works in 4 ways:
* unfiltered: All metrics are reported;
* excludes-only: All metrics are reported, except those whose name
is listed in excludes;
* includes-only: Only metrics whose name is listed in includes are
reported.
* mixed (both *includes* and *excludes*):
Only metrics whose name is listed in includes and not listed in
excludes are reported; excludes takes precedence over includes.