Set a Logback appender encoder rather than layout
Close #882, Close #902

Logback starting from the version 0.9.19 no longer admits a layout
as a component and requires setting an encoder. Currently Logback
adds a warning when a layout is set and internally wraps it as an
encoder.

We should not confuse users with such a warning and make the right
configuration on the framework level.