Ensuring that JerseyRequest scoped ClientConfig gets propagated to HttpUriRequest
**Motivation**

JerseyClient and the ApacheHttpClient are statically configured at creation time. Dropwizard configuration builders ensure those configs are in sync. However, WebTarget objects can have the configuration overridden per generated request. This information will be added to the JerseyRequest. Unfortunately, the current implementation of the DropwizardApacheConnector does not propagate this onto the RequestConfig in the created HttpUriRequest.

In DW 0.7.1 this was working thanks to this code: http://grepcode.com/file/repo1.maven.org/maven2/com.sun.jersey.contribs/jersey-apache-client4/1.18.1/com/sun/jersey/client/apache4/ApacheHttpClient4Handler.java#241

This is currently blocking the migration of https://github.com/yammer/tenacity project to DW 0.8.0, as we need this functionality to sync Hystrix timeouts with those of the HttpClient.

**Solution**

The solution is to update the DropwizardApacheConnector to check the JerseyRequest#getConfiguration for properties (present only if overridden) that should be propagated to the HttpUriRequest specific RequestConfig instance.

I didn't want to loose the default values for not overridden properties, so I wrapped the CloseableHttpClient in a DTO object that bundles together the constructed client and default config used. This detail is package protected (as per code review feedback) and only used internally to construct the `DropwizardApacheConnector`. Other clients still use the original API of the `HttpClientBuilder`.

**Testing**
I've added an integration test that behaviourally tests that the config overrides take effect. This forced me to put a dependency `dropwizard-testing` in test scope in the `dropwizard-client` artifact. I hope this is OK.

Conflicts:
	dropwizard-client/src/test/java/io/dropwizard/client/HttpClientBuilderTest.java