Run tests in parallel
We added a workaround for a bug in sl4fj and removed
clashed tests in Dropwizard modules, so now it's should
be safe to run the tests in parallel.

It could highlight more bugs in tests, because the test
environment now will become more unstable and unpredicatable.
Which is probably not a bad thing, because more bugs will be
found in the development and testing, but in the production.

* Spin-loop the therad while the logger context is not initialized

Because of the bug http://jira.qos.ch/browse/SLF4J-167 a thread
that didn't start logging initialization has a possibilty to get
a reference not to a real logger context, but to a wrapper.

To workaround this we spin-loop the thread, while the context
is not initialized.

* Add MBean registration lock

We have a data race, if LoggerFactory is created in parallel.
Therefore we should guard the process of MBean registration

* Bootstrap the logging before testing

Doing so, we guarantee that the logger context is initialized
during tests.

* Remove not a relevant test

The test that checks that the final frame is the main method
is not relevant, because if it's called in the JUnit parallel
test runner, the last frame is from a custom thread. Because
of this the test fails.

* Use a random server port in Jersey tests

Otherwise these tests start testing HTTP servers on the
same port and they fail when the tests are run in parallel.

* Use random HTTPS port

Othwerise this test may confict with other test, that could
be run on the same port

* Support for a custom property prefix

It's crucial if we don't want situations, when independent
tests concurrently modify the same system property.

* Use different system property prefixes to avoid test clashing

Tests are now running in parallel, so we should use different
system properties for configuration overriding tests, otherwise
they will clash with each other make themselves not reliable.