Run tests in parallel
We added a workaround for a bug in sl4fj and removed
clashed tests in Dropwizard modules, so now it's should
be safe to run the tests in parallel.

It could highlight more bugs in tests, because the test
environment now will become more unstable and unpredicatable.
Which is probably not a bad thing, because more bugs will be
found in the development and testing, but in the production.