Use a custom mustache resolver then a custom factory
A more canonical way to provide a custom resource
loader is through delegation, rather then inheritance.

There is the MustacheResolver interface, that gives ability to
provide own strategy of loading resources and pass it to
the the DefaultMustacheFactory class in the constructor.

Also this removes throwing a MustacheException, because
internally Mustache already handles the case when a returned
reader is null and it throws a MustacheNotFoundException
with the same semantic.