Create managed instances of Apache HTTP Client
In the documentation we say that we create managed instances of
Apache HTTP client. But in the fact, it's not true.

We should tie the lifecyle of the HTTP client to the lifecycle of
the server, if the user specified the environment in the builder.

If the user has already created own managed object - it's not:w
an issue, because the close method of the HTTP client is idempotent.