Add tests for DbDumpCommand
This suite checks 4 cases:

* Dump the existed database schema to an XML change set
* Dump the schema and data to an XML change set
* Dump only data
* Print a correct help page

For checking correctness of a change set, it uses a DOM XML parser,
rather than regular expressions or string manipulations. The reason
for this is that the resulted XML text is not a constant and depends
on environment and time of a run. To overcome this, we use the
XML parser, because it provides ability to make an intelligent
assertions. In the end, it makes tests more stable and easy to
maintain.