Use our hibernate validator in jersey validations
This commit enables `Optional` and `AbstractParam` parameters to be
validated in resource endpoints. This required additional code needed in
setting up the validator, so I extracted the common functionality out to a
new class `Hibernate` that should be analogous to the present `Jackson`
clsas.

Also this commit makes the `NonEmptyStringParam` constructor public like
other `AbstractParam`s

The one area where this commit falls short is that it currently doesn't
allow a user custom hiberate configuration in the jersey validation. Thus,
request entities will be validated by the user's validator but parameters
will not (they'll still be validated by the Dropwizard configured
validator). Since I wasn't sure how best to incorporate this
functionality, I decided to leave it out of this commit.

This allows for some interesting endpoints:

```java
    @GET
    @Path("search")
    public String search(
            @QueryParam("limit")
            @DefaultValue("10")
            @UnwrapValidatedValue
            @Min(1) @Max(100)
            IntParam limit,

            @QueryParam("offset")
            @DefaultValue("0")
            @UnwrapValidatedValue
            @Min(0)
            IntParam offset) {
        // ...
    }
```

It's annotation overload (I believe `@UnwrapValidatedValue` will be
redundant in Hibernate 5.2), but it certainly does pack a lot of
functionality in the parameters.