Simplify authorization
The current implementation of authorization is too cumbersome
and requires implementing a big amount of boilerplate code.

This changes inroduce `Authorizer` interface, which is responsible
for authorizing users. It's a replacement for specifiying a function
for creating a new secure context.

Advantage is stem from the fact that users only implement business
logic for authorization and nothing more. Also this type of API
is similar to `Authenticator`, which makes API more consistent.