Add toString for configuration objects
Objects which hold configuration are serialized from json/yaml files.
Without proper ```toString()``` methods, it is impossible to log what
is the configuration of the application after it was deserialized into
the configuration class. It is especially important in regards to the
```DefaultServerFactory``` class, which holds the configuration for the
```server``` command. By adding the ```toString()``` methods, it is
possible to log during application startup what is the actual configuration
used by the app.