Remove Authenticator generic parameter from AuthFilterBuilder
Generic parameter that controls a type of a passed authenticator
to the `AuthFilterBuilder` is not actually needed. Type-safety is already
forced by the generic parameters C(credentials) and P(principal).
The user has to provide a class that implements interface
`Authenticator<C,P>`, and we don't care about concrete type of
this implementation, because it doesn't affect the auth filter.

This commit also moves the common logic of building auth filters to
the `AuthFilterBuilder` and adds some comments for the public API.