Correct behaviour of the @PermitAll annotation
The `@PermitAll` annotation is actually intended  for marking that a
resource doesn't require authorization. Currently we treat it like the
resource doesn't require authentication.

The solution is to register an auth filter for the methods marked with
the `@PermitAll` annotation. The methods will require authentication,
but anyone, who was given access, can use it.

This simplifies the work of users who don't care about roles and
authorization. They don't need create a fiction role and register
`RolesAllowedDynamicFeature`.