permit extra exceptions to be caught in tests
There appears to be different Exceptions that are thrown depending on the OS
when some of the SSL tests are run. In some cases we may see a
SSLHandshakeException, while in others we see a SocketException due to a
connection reset. I believe this is due to the server's handling of negotiating
TLS sessions different between OSes.

To overcome this, the tests now expect either the SSLHandshakeException or
SocketException in the tests.