Don't close the client instance in DropwizardApacheConnector
Jersey expects that we create a new instance of the Apache HTTP client
every time, when we create a new instance of the connector.

So, when `ClientRuntime` is garbage collected, the client is closed
and new requests, that work with a new connector, use already closed client.

The fix is to not close the client in the connector. We can rely on the fact,
that the client is managed by Dropwizard environment and it will be closed,
when the application will be shut down.

Also users of Apache HTTP Client keep the ability to close the client manually,
if they want to.

Fix #1160