Don't require username in the data source configuration
Some platforms provide credentials to connect to a database
inlined in the connection URL. Examples:

* `postgres://postgres:SOME_PASSWORD@172.17.0.1:5432/lolipop`
* ` jdbc:h2:file:~/sample;USER=sa;PASSWORD=123`
* `jdbc:postgresql://localhost/test?user=fred&password=secret`

JDBC drivers support such a type of configuration. We should
respect it and provide users ability to use it.

It means we should not require the username field to be set in the
data source configuration.

Fixes #1259