Add support for several Hibernate bundles
In some cases one need to simultaneously connect to several
relational databases. Users want to use Hibernate to access each
database. The problem that the current implementation of
`UnitOfWorkApplicationListener` supports only one session factory,
which ultimately restricts ability to support several session factories.

Another problem is that users should have the ability to specify the
name of datasource, against which a transaction should be opened
in a resource method.

This change:

* Adds the method `value` to the `@UnitOfWork`
annotation. It allows to specify the name of Hibernate bundle
(session factory). By default it specifies the default bundle.

* Replaces the single session factory in
`UnitOfWorkApplicationListener` to a map of session factories by
names of bundles.

* The correct session factory is determined in runtime, based on
the `value` method in the corresponding `@UnitOfWork` annotation.

In the result users can simultaneously use declarative transactions
on several databases, access to which is managed by Hibernate.

Fixes #1217