Add support for configuration of HTTP/2 server push.
HTTP/2 supports the server push technology. It affords to a server to
push to a client several additional resources along with a primary
resource. It's very beneficial for dynamic web pages that consists
of many resources apart from an HTML markup (JS libraries, CSS rules,
images, etc).

Jetty provides `ServerPushFilter`, which separates resources on
primary and secondary based on the `Referer` header. The referred
resource within the specified time frame are considered as secondary
to the referer. On the next request to the primary resource secondary
resources will be pushed on a client.