Entities that can't be serialized return a 500
Jackson is good about serializing types that conform to the Bean spec, but
needs help serializing and deserializing types that don't (normally with
the use of annotations). Previous behavior would return a 400 (Bad
Request) when the server would attempt to return an entity that couldn't
be serialized. This is misleading because it gives the illusion that the
fault lies with the client instead of the server.