Add ConfigurationParsingException tests
Add tests for handling errors during parsing and binding YAML configuration.

* Add a test that we print the `Did you mean?` suggestions on an
unrecognized field in the configuration class. The suggestions are
sorted based on an edit distance from the input. The limit is 5 suggestions.
* Add a test that we print a detailed error message, when there are type
data type errors during binding a YAML tree to the configuration class.
* Add a test that we print a line number, a position of an error, and a
parser error message, if a YAML file is malformed.