ValidationMethod violation message is itself
Previously, when a ValidationMethod caused a contraint violation and the
violation was caused by a nested object using the `@Valid` annotation, the
message would contain the path of the nested object. This may be
misleading to the client who sees the message because the nested code path
may not have a related field in the JSON they sent (for example, the field
could be a misnomer).

The fix is return the message exactly as it is transcribed in the
`@ValidationMethod` annotation. Thus, it is now the user's responsibility
to create descriptive enough messages, such that the client can discern
the error through any amount of nesting.

This change does APIs that relied on the message containing the code path
to the object that contained the ValidationMethod violation.