Fix misleading maxThreads default & misuse log warning
When maximumPoolSize > corePoolSize with a unbound work queue
(ie the default LinkedBlockingQueue) the maximumPoolSize setting
has no effect. The max threads the pool uses is the corePoolSize
(or 1 if the corePoolSize is 0). This means the current warning
logic isn't quite there, and worse the default implementation
is misleading and actually results in a 1-thread pool.

To change the behaviour the least I've changed the default
maxThreads to 1. Now if manually set to >1 but with an
unbound work queue we'll log a warning. (A warning existed
before but wasn't triggering for MAX_VALUE maxThreads)

Furthermore, the warning was also appearing for other legitimate
usages like `.minThreads(1).maxThreads(1)`. The warning should
now only appear when a misleading maxThreads has been used.
Really we should consider throwing here, or changing the API
to something more foolproof like the Executors methods.