configures the context paths before the application.run()
This commit configures the context paths for the application before the
application `run()` instead of setting it prior to starting the server.
This allows applications to obtain the context paths during in the run
method.

This shouldn't break any changes to the users, but note that
ServerFactory now needs to call `configure` explicitly to set the
context paths in the environment