Enhancements to lifecycle-setup classes.
- Added constructors for executor service builders to allow specifying a
thread factory so that a thread factory is not built only to be thrown
away by a later call to the method to set the thread factory; the
ability to override the thread factory has not been removed for backward
compatibility.
- Exposed factory methods that create the builders based on already
existing thread factories, thus invoking the new constructors created in
the above comment.