Document the Dropwizard release process
Add a simple document, which describes the technical aspects of making a release of Dropwizard.
This information is mostly exist in the form emails between Dropwizard commiters and personal
experience, so it's quite easy to a maintainer to forget some steps during a release and make a
mistake. It would be great, if we wrote it down in the repository and make accessible to the community.
This should make Dropwizard a more transparent project and allow to improve our release process.

[ciskip]

Use script