Add a test for building the HTTPS connector
I noticed that we don't have a test for building the HTTPS connector in `HttpsConnectionFactory`.
There is a lot of things are going on during this process (building a SSL context factory, logging
supported ciphers, setting metric names, using correct instrumented connectors and so on),
so it makes sense to test it independently from other components. I think it would be great to
have a test which verifies that we correctly configure Jetty on the SSL side.

As a result, this should allow to spot possible bugs earlier during future Jetty updates.