Expose the request post body to tasks.
Following the example of the task query params, the body is just exposed
as a raw string. The task of parsing the body into some useful entity is
left to the user for maximum flexibility.

As of now, this is a breaking change, because any existing task
subclassed will need to be updated. I may change that depending on
feedback from the Dropwizard team.