Commit a database transaction in the `RESP_FILTERS_START` stage
But close the session during the `FINISHED` stage. This allows
to the user to catch errors during commiting transactions
(constraint violations, e.g) and map them to HTTP error with
an exception mapper.

This approach also is compatible with lazy loading, because
the session remain open during serializing Hibernate entities.

Fix #1604