Set FAIL_ON_UNKNOWN_PROPERTIES in DefaultConfigurationFactoryFactory (#1677)
ObjectMapper-customization logic is moved from YamlConfigurationFactory
to DefaultConfigurationFactoryFactory, making it easier to modify with a
custom ConfigurationFactoryFactory implementation.

There are some (uncommon) use cases where it's beneficial to update an
application without modifying the configuration, which may fail if the
application configuration has removed a configuration property.

A protected configureObjectMapper method is added to
DefaultConfigurationFactoryFactory to allow the configuration
factory to be extended to apply (or remove) deserialization
features.